requirements.txt
typer
pymongo
python-dotenv